iox-#1032 Rename platform to custom
