+++ iceoryx_hoofs/include/iceoryx_hoofs/error_reporting/custom/default/configuration.hpp
@@ -27,7 +27,7 @@ namespace err
 
 // Specialize to change the checks (and other options if needed)at compile time.
 // this can later also be done depending on a #define to select a header
-// but we should avoid to have a define for each option.
+// but we should avoid to have a #define for each option.
 template <>
 struct ConfigurationParameters<ConfigurationTag>
 {
