+++ iceoryx_hoofs/include/iceoryx_hoofs/error_reporting/custom/default/error_handler_interface.hpp
@@ -14,11 +14,11 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef IOX_HOOFS_ERROR_REPORTING_PLATFORM_ERROR_HANDLER_INTERFACE_HPP
-#define IOX_HOOFS_ERROR_REPORTING_PLATFORM_ERROR_HANDLER_INTERFACE_HPP
+#ifndef IOX_HOOFS_ERROR_REPORTING_CUSTOM_ERROR_HANDLER_INTERFACE_HPP
+#define IOX_HOOFS_ERROR_REPORTING_CUSTOM_ERROR_HANDLER_INTERFACE_HPP
 
 #include "iceoryx_hoofs/error_reporting/errors.hpp"
-#include "iceoryx_hoofs/error_reporting/location.hpp"
+#include "iceoryx_hoofs/error_reporting/source_location.hpp"
 #include "iceoryx_hoofs/error_reporting/types.hpp"
 
 namespace iox
