--- iceoryx_hoofs/source/error_reporting/platform/default/default_error_handler.cpp
@@ -1,21 +1,0 @@
-#include "iceoryx_hoofs/error_reporting/platform/default/default_error_handler.hpp"
-
-namespace iox
-{
-namespace err
-{
-
-void DefaultHandler::reportError(ErrorDescriptor)
-{
-}
-
-void DefaultHandler::reportViolation(ErrorDescriptor)
-{
-}
-
-void DefaultHandler::panic()
-{
-}
-
-} // namespace err
-} // namespace iox
