+++ iceoryx_hoofs/test/moduletests/error_reporting/test_test_error_handler.cpp
@@ -19,7 +19,7 @@
 #include "test.hpp"
 #include <gtest/gtest.h>
 
-#include "iceoryx_hoofs/error_reporting/location.hpp"
+#include "iceoryx_hoofs/error_reporting/source_location.hpp"
 #include "iceoryx_hoofs/testing/error_reporting/test_error_handler.hpp"
 
 // NOLINTNEXTLINE(hicpp-deprecated-headers) required to work on some platforms
